[Sign Up] Add avatar uploading to signUpUser action
